Change TableRowColAuthResult to TableRowColAccessResult
