+++ pinot-broker/src/main/java/org/apache/pinot/broker/api/AccessControl.java
@@ -26,8 +26,8 @@ import org.apache.pinot.spi.annotations.InterfaceStability;
 import org.apache.pinot.spi.auth.AuthorizationResult;
 import org.apache.pinot.spi.auth.BasicAuthorizationResultImpl;
 import org.apache.pinot.spi.auth.TableAuthorizationResult;
-import org.apache.pinot.spi.auth.TableRowColAuthResult;
-import org.apache.pinot.spi.auth.TableRowColAuthResultImpl;
+import org.apache.pinot.spi.auth.TableRowColAccessResult;
+import org.apache.pinot.spi.auth.TableRowColAccessResultImpl;
 import org.apache.pinot.spi.auth.broker.RequesterIdentity;
 
 
@@ -128,9 +128,9 @@ public interface AccessControl extends FineGrainedAccessControl {
    * Returns RLS/CLS filters for a particular table. By default, there are no RLS/CLS filters on any table.
    * @param requesterIdentity requested identity
    * @param table Table used in the query. Table name can be with or without tableType.
-   * @return {@link TableRowColAuthResult} with the result of the access control check
+   * @return {@link TableRowColAccessResult} with the result of the access control check
    */
-  default TableRowColAuthResult getRowColFilters(RequesterIdentity requesterIdentity, String table) {
-    return TableRowColAuthResultImpl.unrestricted();
+  default TableRowColAccessResult getRowColFilters(RequesterIdentity requesterIdentity, String table) {
+    return TableRowColAccessResultImpl.unrestricted();
   }
 }
