+++ pinot-spi/src/main/java/org/apache/pinot/spi/auth/TableRowColAccessResult.java
@@ -22,30 +22,15 @@ import java.util.List;
 import java.util.Optional;
 
 
-public class TableRowColAuthResultImpl implements TableRowColAuthResult {
+/**
+ * This interfaces carries the RLS/CLS filter for a particular table
+ */
+public interface TableRowColAccessResult {
+  /**
+   * Returns the RLS filters associated with a particular table. RLS filters are defined as a list.
+   * @return optional of the RLS filters. Empty optional if there are no RLS filters defined on this table
+   */
+  Optional<List<String>> getRLSFilters();
 
-  private static final TableRowColAuthResult UNRESTRICTED = new TableRowColAuthResultImpl();
-
-  private List<String> _rlsFilters;
-
-  public TableRowColAuthResultImpl() {
-  }
-
-  public TableRowColAuthResultImpl(List<String> rlsFilters) {
-    _rlsFilters = rlsFilters;
-  }
-
-  @Override
-  public void setRLSFilters(List<String> rlsFilters) {
-    _rlsFilters = rlsFilters;
-  }
-
-  @Override
-  public Optional<List<String>> getRLSFilters() {
-    return Optional.ofNullable(_rlsFilters);
-  }
-
-  public static TableRowColAuthResult unrestricted() {
-    return UNRESTRICTED;
-  }
+  void setRLSFilters(List<String> rlsFilters);
 }
